requirements.txt
PyAissistant~=0.1.17
openai
ollama-api~=0.1.1
requests~=2.32.3
watchdog~=4.0.1
colorama~=0.4.6

colorlog~=6.8.2

elasticsearch~=8.15.1

PyYAML~=6.0.1
